requirements.txt
openai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
httpx>=0.25.0
ddgs>=4.0.0
pymupdf>=1.22.0
pymupdf4llm>=0.0.8
python-multipart>=0.0.6
python-dotenv>=1.0.0
PyYAML>=6.0.0
nest-asyncio>=1.5.0
python-socks>=2.8.0
python-docx>=1.1.0
fpdf2>=2.7.0
openpyxl>=3.1.0
pandas>=2.0.0
tabulate>=0.9.0
langfuse>=2.0.0
docker>=6.1.0
browser-use
MarkItDown[all]
typer>=0.9.0
litellm>=1.0.0
loguru>=0.7.0
python-telegram-bot>=21.0
lark-oapi>=1.2.0
rich>=13.0.0
pydantic-settings>=2.0.0
questionary>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
apscheduler>=3.10.0
arxiv>=2.0.0
scholarly>=1.7.11
semanticscholar>=0.7.0
beautifulsoup4>=4.11.0
pyoverleaf
GitPython>=3.1.0
apprise>=1.9.0
overleaf-sync>=1.2.0
dingtalk-stream>=0.24.3
# Add any other specific dependencies for context_bot if needed
